Failing tests for selecting certain things
testSelectionState queries the style state of a selection, and expects
Mozilla's behaviour (ie, if parts of the selection have italics and
other parts don't, querying the italics style on that selection
should return false). This test fails in Safari.

testSelectImage is a test for determining if an image has been selected.
It currently fails in Firefox and Safari.

Signed-off-by: Joshua Peek <josh@joshpeek.com>